feat: linting rule for unregistered external models
[ci skip]
